+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Autoscaler Change Log
 ------------------------
 
 - Fix MaxListenersExceededWarning in logger
+- Fix number of machines included in scaling loop, use '<= MAX'
 
 0.6.0 October 28, 2020
 ------------------------